requirements.txt
google-generativeai
python-dotenv
ruff
pillow